Two conflicts, both where #10332's macOS speech backend landed in the same
insertion points this branch uses:

- src/CMakeLists.txt: kept both sides. AppleSpeechRecognizer.mm stays in the
  if(APPLE) branch with the uiawrapper.cpp elseif(WIN32) branch after it, and
  the two header appends each keep their own platform guard.
- src/mudlet.cpp: upstream removed the system physical memory include block
  and the libzip version check outright, so the only part of this branch's
  side still wanted is the uiawrapper.h include, kept inside the existing
  Q_OS_WINDOWS guard.

+++ src/TAreaGridIndex.cpp
@@ -26,12 +26,15 @@ const TAreaGridIndex::RoomIds TAreaGridIndex::csmEmptyCell;
 // it. A zoomed-in viewport covers a handful of the rows a level occupies and a
 // zoomed-out one covers more rows than exist, so neither wins outright - the
 // smaller key count is the one to get through.
+//
+// The probing loops count in qint64 because an int counter at INT_MAX overflows
+// back to INT_MIN rather than passing a bound of INT_MAX, and never terminates.
 static void appendRoomsInYRange(const QHash<int, TAreaGridIndex::RoomIds>& yMap, int minY, int maxY, QList<int>& result)
 {
     const qint64 wantedRows = static_cast<qint64>(maxY) - minY + 1;
     if (wantedRows > 0 && wantedRows < yMap.size()) {
-        for (int y = minY; y <= maxY; ++y) {
-            const auto yIt = yMap.constFind(y);
+        for (qint64 y = minY; y <= maxY; ++y) {
+            const auto yIt = yMap.constFind(static_cast<int>(y));
             if (yIt == yMap.constEnd()) {
                 continue;
             }
@@ -55,8 +58,8 @@ static void appendRoomsInYRangeWithCollisions(const QHash<int, TAreaGridIndex::R
 {
     const qint64 wantedRows = static_cast<qint64>(maxY) - minY + 1;
     if (wantedRows > 0 && wantedRows < yMap.size()) {
-        for (int y = minY; y <= maxY; ++y) {
-            const auto yIt = yMap.constFind(y);
+        for (qint64 y = minY; y <= maxY; ++y) {
+            const auto yIt = yMap.constFind(static_cast<int>(y));
             if (yIt == yMap.constEnd()) {
                 continue;
             }
@@ -190,8 +193,8 @@ QList<int> TAreaGridIndex::roomsInViewport(int z, int minX, int maxX, int minY, 
     const auto& xyMap = *zIt;
     const qint64 wantedColumns = static_cast<qint64>(maxX) - minX + 1;
     if (wantedColumns > 0 && wantedColumns < xyMap.size()) {
-        for (int x = minX; x <= maxX; ++x) {
-            const auto xIt = xyMap.constFind(x);
+        for (qint64 x = minX; x <= maxX; ++x) {
+            const auto xIt = xyMap.constFind(static_cast<int>(x));
             if (xIt == xyMap.constEnd()) {
                 continue;
             }
@@ -218,8 +221,8 @@ QList<QPair<int, bool>> TAreaGridIndex::roomsInViewportWithCollisions(int z, int
     const auto& xyMap = *zIt;
     const qint64 wantedColumns = static_cast<qint64>(maxX) - minX + 1;
     if (wantedColumns > 0 && wantedColumns < xyMap.size()) {
-        for (int x = minX; x <= maxX; ++x) {
-            const auto xIt = xyMap.constFind(x);
+        for (qint64 x = minX; x <= maxX; ++x) {
+            const auto xIt = xyMap.constFind(static_cast<int>(x));
             if (xIt == xyMap.constEnd()) {
                 continue;
             }
